requirements.txt
# FastAPI and web server dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Image processing dependencies
opencv-python>=4.8.0
scikit-image>=0.20.0
Pillow>=10.0.0
numpy>=1.21.0

# HTTP client for downloading images
requests>=2.28.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Caching dependencies
redis>=5.0.0

# Additional utilities
python-dotenv>=1.0.0

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
